Étape 2 - modules CheckElf
