+++ modules/CheckElf.c
@@ -1,0 +1,23 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+int checkELF(FILE *file){
+	char buffer[4];
+	fread(buffer, 4, 1, file);
+	if (feof(file)){
+		return 0;
+	}
+	if (buffer[0] != 0x7F){
+		return 0;
+	}
+	if (buffer[1] != 0x45){
+		return 0;
+	}
+	if (buffer[2] != 0x4C){
+		return 0;
+	}
+	if (buffer[3] != 0x46){
+		return 0;
+	}
+	return 1;
+}
