+++ modules/CheckElf.h
@@ -1,0 +1,5 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+/* Check if a file is of type ELF*/
+int checkELF(FILE *file);
