+++ read_elf_header.c
@@ -2,29 +2,7 @@
 #include <stdlib.h>
 #include "modules/freadoctet.h"
 #include "modules/CustomElf.h"
-
-
-int checkELF(FILE *file){
-	char buffer[4];
-	fread(buffer, 4, 1, file);
-	if (feof(file)){
-		return 0;
-	}
-	if (buffer[0] != 0x7F){
-		return 0;
-	}
-	if (buffer[1] != 0x45){
-		return 0;
-	}
-	if (buffer[2] != 0x4C){
-		return 0;
-	}
-	if (buffer[3] != 0x46){
-		return 0;
-	}
-	return 1;
-}
-
+#include "modules/CheckElf.h"
 
 void magicNumber(FILE *file, Elf32_Ehdr* Header, char verbose){
 	if (verbose){
